Cpg134 product helper murugan

See merge request IvyProjects/IvyCpg_Product_MobileApp_Android!475

+++ ivycpg/src/main/java/com/ivy/sd/png/provider/ProductHelper.java
@@ -1103,7 +1103,7 @@ public class ProductHelper {
 
                     product.setPriceoffvalue(c.getDouble(c.getColumnIndex("priceoffvalue")));
                     product.setPriceOffId(c.getInt(c.getColumnIndex("priceoffid")));
-                    product.setASRP(c.getInt(c.getColumnIndex("asrp"))); //added by murugan
+                    product.setASRP(c.getFloat(c.getColumnIndex("asrp"))); //added by murugan
 
                     product.setAvailableinWareHouse(c.getString(c.getColumnIndex("IsAvailWareHouse")).equals("true"));
                     product.setHsnId(c.getInt(c.getColumnIndex("HSNId")));
